[chore] #33 DSKit에 있는 protocol Util로 이동
