+++ Projects/Util/Sources/Protocols/PokitSelectItem.swift
@@ -1,0 +1,13 @@
+//
+//  PokitSelectItem.swift
+//  Util
+//
+//  Created by 김도형 on 7/11/24.
+//
+
+import Foundation
+
+public protocol PokitSelectItem: Identifiable, Equatable {
+    var categoryType: String { get }
+    var contentSize: Int { get }
+}
